Refactor test examples and update method documentation
Moved some test examples to a new `example_slice_test.go` file and removed them from the `examples_test.go` file. Updated documentation for the `Every` and `Filter` methods in `godash.go` to be more narrowly focused on their method behaviors, including usage examples.
